requirements.txt
amqp==5.2.0
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.8.2
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
dj-database-url==2.2.0
Django==5.1.1
django-celery-beat==2.7.0
django-cors-headers==4.5.0
django-environ==0.11.2
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
eventlet==0.37.0
Flask==3.0.3
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
MarkupSafe==3.0.1
MouseInfo==0.1.3
pillow==10.4.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyJWT==2.9.0
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2024.2
redis==5.1.1
requests==2.32.3
setuptools==75.2.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.4
wikipedia==1.4.0
